Change drive_prep() to look for the device instead of the mount
There is a subtle but annoying bug where using tab completion to give
the mount path adds a "/" at the end. When `grep` is run to check
if the image is mounted it searches with the trailing slash.

By looking for the partition instead we can avoid this small usability
issue which can cause confusion.